Rename namespace from 'Slack.Webhooks.SlackAction' to 'Slack.Webhooks.Action'. Rename class from Action to SlackAction
