+++ src/Slack.Webhooks/Action/OptionGroup.cs
@@ -1,6 +1,6 @@
 using System.Collections.Generic;
 
-namespace Slack.Webhooks.SlackAction
+namespace Slack.Webhooks.Action
 {
     public class OptionGroup
     {
