+++ src/Slack.Webhooks/SlackAttachment.cs
@@ -77,6 +77,6 @@ namespace Slack.Webhooks
         /// <summary>
         /// The actions you provide will be rendered as message buttons or menus to users.
         /// </summary>
-        public List<SlackAction.Action> Actions { get; set; }
+        public List<Action.SlackAction> Actions { get; set; }
     }
 }
